+++ examples/apps/lit-router/src/store/index.js
@@ -1,14 +1,14 @@
-import { createStore } from "@inglorious/lit"
+import { createStore, router } from "@inglorious/lit"
 
 import { entities } from "./entities"
 import { middlewares } from "./middlewares"
-import { router } from "./types/router"
 import { home } from "./types/home"
 import { userList } from "./types/user-list"
 import { userDetail } from "./types/user-detail"
+import { notFound } from "./types/not-found"
 
 export const store = createStore({
-  types: { router, home, userList, userDetail },
+  types: { router, home, userList, userDetail, notFound },
   entities,
   middlewares,
 })