+++ examples/apps/lit-router/src/store/types/not-found.js
@@ -1,0 +1,8 @@
+import { html } from "@inglorious/lit"
+
+export const notFound = {
+  render() {
+    return html`<h1>Uh-oh...</h1>
+      <p>That's embarassing. Please look somewhere else.</p>`
+  },
+}