+++ examples/apps/lit-router/CHANGELOG.md
@@ -1,5 +1,16 @@
 # lit-todomvc
 
+## 0.2.0
+
+### Minor Changes
+
+- Add router type to @inglorious/lit
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/lit@1.1.0
+
 ## 0.1.1
 
 ### Patch Changes