+++ examples/apps/lit-todomvc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lit-todomvc
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/lit@1.1.0
+
 ## 0.1.1
 
 ### Patch Changes